Removed the need for Delay Time on Borderlands 2

+++ Master/NucleusGaming/Interop/User32.cs
@@ -10,9 +10,9 @@ namespace Nucleus.Gaming.Interop
     {
         public static void HideBorder(IntPtr handle)
         {
-            uint lStyle = (uint)User32Interop.GetWindowLong(handle, User32_WS.GWL_STYLE);
+            uint lStyle = User32Interop.GetWindowLong(handle, User32_WS.GWL_STYLE);
             lStyle &= ~(User32_WS.WS_CAPTION | User32_WS.WS_BORDER | User32_WS.WS_DLGFRAME | User32_WS.WS_SIZEBOX | User32_WS.WS_THICKFRAME);
-            User32Interop.SetWindowLong(handle, User32_WS.GWL_STYLE, (int)lStyle);
+            User32Interop.SetWindowLong(handle, User32_WS.GWL_STYLE, lStyle);
         }
         public static void HideTaskbar()
         {
